Added Player Health, Speed States, minor tweeks
Added several int references to include PlayerHealth and threshold points (Full, Injured, Half, Critical) as well as GameEvents and listeners for when the player's health changes and if it is equal to one of these thresholds. The PlayerHealthObserver script checks each time the player's health changes and will fire events accordingly based on if it is equal to one of the thresholds, if the player gained/lost health, and if the player is dead.

Added Medkit prefab, script, and int reference which restores player's health and destroys itself on contact with the player.

Added float references for player speed and connected them to the PlayerMovement script, as well as implementing the speed change when the Shift/Run key is pressed. Added a blend tree for the player Animator that will play correct animation based on speed, horizontal, and vertical movement. A little difficult to test it fully since the Morgue is fairly cramped at the moment.

Suppressed 14 warnings which were appearing because the "Developer Notes" on some scriptable objects "weren't being used"